requirements.txt
hikari-py
hikari-tanjun
python-dotenv
